+++ src/app/services/users.service.ts
@@ -922,14 +922,32 @@ export class UsersService {
 
     // const body = { 'role': user_role, 'max_served_chat': max_served_chat };
     const body = { 'role': user_role, 'max_assigned_chat': max_assigned_chat };
-
-
     this.logger.log('[USER-SERV] - UPDATE PROJECT-USER ROLE & MAX-CHAT  BODY ', body);
-
     return this.http
       .put(url, JSON.stringify(body), options)
       .map((res) => res.json());
   }
+
+  // / curl -v -X PUT -H 'Content-Type:application/json' -u andrea.leo@f21.it:123456 -d '{"tags":[{"tag": "inprogress", "color": "#66C549"}]}'
+  // //  http://localhost:3001/6256ac8c729977ad37f0aee6/project_users/ID_PROJECT_USER
+  public updateProjectUserTags(projectUser_id: string, tagarray: any) {
+    let url = this.PROJECT_USER_URL + projectUser_id;
+    console.log('[USER-SERV] - UPDATE PROJECT-USER TAG URL ', url);
+
+    const headers = new Headers();
+    headers.append('Accept', 'application/json');
+    headers.append('Content-type', 'application/json');
+    headers.append('Authorization', this.TOKEN);
+    const options = new RequestOptions({ headers });
+
+    // const body = { 'role': user_role, 'max_served_chat': max_served_chat };
+    const body = { 'tags': tagarray };
+    console.log('[USER-SERV] - UPDATE PROJECT-USER TAG  BODY ', body);
+    return this.http
+      .put(url, JSON.stringify(body), options)
+      .map((res) => res.json());
+  }
+
 
 
   /**